Add go-socks5-proxy dependency for CI and test with auth
