+++ .github/workflows/test.yml
@@ -13,6 +13,17 @@ permissions:
 jobs:
   test:
     runs-on: ubuntu-latest
+
+    services:
+      services:
+        socks5:
+          image: serjs/go-socks5-proxy
+          env:
+            PROXY_USER: user
+            PROXY_PASSWORD: password
+          ports:
+            - 1080:1080
+
     strategy:
       matrix:
         ruby: ['3.1', '3.2', '3.3', head]
