+++ ChangeLog
@@ -82,4 +82,4 @@ SOCKSify Ruby 1.7.3
 * add Rakefile
 * fix missing :timeout kwarg in TCPSocket class (thanks @lizzypy)
 * Authentication support added to Net::HTTP.SOCKSProxy
-  (thanks to @ojab)
+  (thanks to @ojab and @anton-smagin)
